+++ phillydb/utils.py
@@ -26,10 +26,14 @@ def get_normalized_address(address):
     combined = " ".join(
         (
             n_address_dict["address_line_1"]
-            + " " + n_address_dict["address_line_2"]
-            + " " + n_address_dict["city"]
-            + " " + n_address_dict["state"]
-            + " " + n_address_dict["postal_code"]
+            + " "
+            + n_address_dict["address_line_2"]
+            + " "
+            + n_address_dict["city"]
+            + " "
+            + n_address_dict["state"]
+            + " "
+            + n_address_dict["postal_code"]
         )
         .strip()
         .split()