+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Development
 -----------
 
-* Placeholder
+* Add market_value to owner query results.
 
 0.2.4
 -----